apple buildfix

+++ platform/sdl2/platform.c
@@ -508,10 +508,10 @@ rate_of_refresh()
   return refresh_rate;
 }
 
-extern int getentropy(void* buffer, size_t size);
-extern ssize_t getrandom(void* buf, size_t buflen, unsigned int flags);
+extern int getentropy(void* buffer, uint64_t size);
+extern int64_t getrandom(void* buf, uint64_t buflen, unsigned int flags);
 extern int open(const char* pathname, int flags);
-extern ssize_t read(int fd, void* buf, size_t count);
+extern int64_t read(int fd, void* buf, uint64_t count);
 extern int close(int fd);
 ptrsize
 platform_random()
